Fix BigDecimal ruby 2.6 deprecation warnings
With the release of Ruby 2.6.0 using `BigDecimal.new` now emits the
following deprecation warning:

```
warning: BigDecimal.new is deprecated; use BigDecimal() method instead.
```

This commit fixes these warnings which will unblock a potential Ruby 2.6.0
upgrade

**Notes**

It is worth noting that these warnings will also be fixed with the
upcoming release of shoulda-matchers `4.0.0` which will remove the need
for this fork.